# api/controllers/movies.js
'use strict';
/*
 'use strict' is not required but helpful for turning syntactical errors into true errors in the program flow
 https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Strict_mode
 */

/*
 Modules make it possible to import JavaScript files into your application.  Modules are imported
 using 'require' statements that give you a reference to the module.

 It is a good idea to list the modules that your application depends on in the package.json in the project root
 */
var util = require('util');

var mongoose = require('mongoose');
var Movie = mongoose.model('Movie');


/*
 Once you 'require' a module you can reference the things that it exports.  These are defined in module.exports.

 For a controller in a127 (which this is) you should export the functions referenced in your Swagger document by name.

 Either:
 - The HTTP Verb of the corresponding operation (get, put, post, delete, etc)
 - Or the operationId associated with the operation in your Swagger document

 In the starter/skeleton project the 'get' operation on the '/hello' path has an operationId named 'hello'.  Here,
 we specify that in the exports of this module that 'hello' maps to the function named 'hello'
 */
module.exports = {
    queryMovies: queryMovies,
    getMovie: getMovie,
    createMovie: createMovie,
    updateMovie:updateMovie,
    deleteMovie: deleteMovie,
    getActors: getActors,
};

/*
 Functions in a127 controllers used for operations should take two parameters:

 Param 1: a handle to the request object
 Param 2: a handle to the response object
 */
function queryMovies(req, res) {
    // variables defined in the Swagger document can be referenced using req.swagger.params.{parameter_name}

    Movie.find(function (err, movies) {
        if (err) {
            return next(err);
        }

        res.json(movies);
    });
}

function getMovie(req, res) {
    // variables defined in the Swagger document can be referenced using req.swagger.params.{parameter_name}

    Movie.findOne({ 'id':req.swagger.params.id.value } ,function (err, movie) {
        if (err) {
            res.status(400).send(err);
        }
        res.json(movie);
    });
}

function createMovie(req, res) {
    // variables defined in the Swagger document can be referenced using req.swagger.params.{parameter_name}

    var movie = Movie(req.swagger.params.movie.value);

    if(!movie._id){
        movie._id = mongoose.Types.ObjectId();
    }

    movie.save(function (err, movie) {
        if (err) {
            res.status(500).send(err);
        }

        res.json(movie);
    });
}


function updateMovie(req, res) {
    // variables defined in the Swagger document can be referenced using req.swagger.params.{parameter_name}

    Movie.update( { _id: req.swagger.params.id.value }, req.swagger.params.movie.value, function (err, movie) {

        if (err) {
            res.status(400).send(err);
        }

        res.json(movie);
    });
}

function deleteMovie(req, res) {
    // variables defined in the Swagger document can be referenced using req.swagger.params.{parameter_name}

    Movie.findByIdAndRemove(req.swagger.params.id.value, function(err) {
        if (err) {
            res.status(400).send(err);
        }

        res.json({message: "Successfully deleted"});
    });
}

function getActors(req, res) {
    // variables defined in the Swagger document can be referenced using req.swagger.params.{parameter_name}

    Movie.findOne({'id': req.swagger.params.id.value}, function (err, movie) {

        if (err) {
            res.status(400).send(err);
        }
        res.json(movie.actors);
    });
}