# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Hello World App
# during dev, should point to your local machine
host: localhost:10010
# basePath prefixes all resource paths 
basePath: /
# 
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /movies:
    x-swagger-router-controller: movies
    get:
      description: Returns all movies
      operationId: queryMovies
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            type: array
            items:
              $ref: "#/definitions/Movie"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      description: Creates a new movie
      operationId: createMovie
      parameters:
        - name: movie
          in: body
          description: The movie which should be created
          required: true
          schema:
            $ref: "#/definitions/Movie"
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/Movie"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /movies/{id}:
    x-swagger-router-controller: movies
    get:
      description: Returns the movie with the specified ID
      operationId: getMovie
      parameters:
        - name: id
          in: path
          description: The ID of the movie
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/Movie"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    put:
      description: Updates the specified movie
      operationId: updateMovie
      parameters:
        - name: id
          in: path
          description: The ID of the movie
          required: true
          type: string
        - name: movie
          in: body
          description: The updated movie
          required: true
          schema:
            $ref: "#/definitions/Movie"
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/Movie"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      description: Deletes the specified movie
      operationId: deleteMovie
      parameters:
        - name: id
          in: path
          description: The ID of the movie to be deleted
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/SuccessResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /movies/{id}/actors:
    x-swagger-router-controller: movies
    get:
      description: Returns all actors for the specified movie
      operationId: getActors
      parameters:
        - name: id
          in: path
          description: The ID of the actor
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            type: array
            items:
              $ref: "#/definitions/Actor"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

  /actors:
    x-swagger-router-controller: actors
    get:
      description: Returns all actors
      operationId: getActors
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            type: array
            items:
              $ref: "#/definitions/Actor"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
# complex objects have schema definitions
definitions:
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
  SuccessResponse:
    properties:
      message:
        type: string
  Movie:
    required:
      - title
    properties:
      title:
        type: string
      description:
        type: string
      year:
        type: number
  Actor:
    required:
      - name
    properties:
      name:
        type: string